Lab 3 Completed
Please take a look at InfilExfilAgent.java located in src/labs/infexf/agents. This agent has
two methods that you need to complete: getEdgeWeight and shouldReplacePlan. Like last time,
the shouldReplacePlan method returns true if the current plan is invalid (for instance if something
blocks you), and false otherwise. My code will use this method to recalculate a plan whenever this
method returns true.
The majority of the work for this lab is the getEdgeWeight method. This method returns the weight
of an edge between two vertices, and is extremely important to the success of the mission. If your edge
weights do not account for “danger” or “risk” (for instance, if an edge brings you within the attack
radius of the enemy, you certainly should discourage A* from using that edge in a path). Basically,
we want to assign edges that are bad large weights (so A* wont use them), and assign edges that are
good small weights (so A* will use them). This will involve quite a bit of creativity on your part,
which is why I have implemented A* (and the state machine of the agent) for you.
When testing, the two smallers maps OneUnitSmallMaze.xml and TwoUnitSmallMaze.xml are always
winnable. For full credit, your agent must win 100% of these games that it plays (I will play a bunch
of these games on the autograder).