Lab 6 Files Uploaded - Pitfall
3. Pitfall Rules & Information
The game of Pitfall is very similar (in fact a simpli ed) version of another game called Minesweeper.
You are given a grid of squares, the contents of each square are initially hidden from you. The player
interacts with the game by clicking on a (currently hidden) square, at which point the contents of
that square are revealed. A square can contain two kinds of things: a pit, or nothing. If you click on
a square that reveals a pit, you fall into the pit and you lose. If you click on a sqaure that does not
have a pit, you are instead told whether or not adjacent squares contain pits (cardinal directions are
adjacent) through the form of a breeze. If you perceive a breeze, you know that some combination of
unknown adjacent squares contain pits.

Any square that has a green box within it are hidden (i.e. unclicked on) squares. Your agent controls
the red \a" at the bottom of the grid. The border of unmarked squares are not part of the game:
they are only there so that the \a" has somewhere to exist on the map.

Your agent will play Pitfall by attacking squares in the map. Your agent is only allowed to attack
hidden squares (i.e. squares with green boxes in them): attacking any other square will result in that
square  ring back at you (and killing you). Your unit as well as all hidden squares die in one hit.
Once a hidden square has been killed, the enemy agent will replace that unit with a new unit whose
ascii-art is either empty (a safe square) or three dotten lines (a breeze).

Your agent needs to use the breeze information (called \clues") to  gure out the probability that each
of the \frontier" squares (the set of unexplored squares adjacent to explored squares) contain a pit
given the breeze information that we have seen. Essentially, I am asking you to use the equation we
derived in class to implement a specific method that will calculate this probability.