Basic Implementation Complete
Methods:
valid_movement(Ship current_Ship, EnemyBoard.Outcome[][] current_board, GameView game):
Checks if a given ship's placement is valid on the current board. It ensures that the ship's coordinates are within bounds and not previously hit (marked as MISS or SUNK).
getBoardOfProb(ShipType ship_Type, EnemyBoard.Outcome[][] current_Board, int board_rows, int board_cols, final GameView game):
Calculates the probability of each cell on the board being the location of a ship of a specific type. It does this by iterating over all possible orientations and positions of the ship, checking for valid placements, and incrementing counters for each cell occupied by these valid placements. The method then normalizes these counts into probabilities by dividing each by the total number of valid placements found.
makeMove(final GameView game):
Determines the best move to make based on the calculated probabilities of the presence of enemy ships on the board. It considers the types and numbers of enemy ships that have not yet been sunk and aggregates the probability boards for each ship type, weighted by the number of ships of that type. The method selects the coordinate with the highest aggregated probability that has not been targeted yet (marked as UNKNOWN) as its next move.

Next i want to create some logic to assign higher probabilties to zones that are hit.