Lab 7 - Zombayes - Naive Bayes Classifier Implementation
In this lab we will be playing a zombie survival game called Zombayes, where we have to classify between human units (pretending to be zombies so they don’t get eaten) and actual zombies. Zombies will attack you given the chance, but humans will not. Your job will be to use a decision tree to learn how to distinguish between zombies and humans: you want to attack the zombies but not the humans!

Removed alot of files due to space constraints, main class remains.